# docs/api.yaml
openapi: 3.0.3
info:
  title: KiloCalorie API
  description: A RESTful API for tracking daily caloric intake
  version: 0.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /users/{id}/days:
    get:
      summary: Get all days for a user
      description: Get all information for a user's days
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The id of the day
                    date:
                      type: string
                      format: date
                      description: The date of the day
        '401':
          description: Unauthorized
        '404':
          description: User does not exist
  /users/{id}/days/{date}:
    get:
      summary: Get a day for a user by date
      description: Get all information for a user's day
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user
          schema:
            type: integer
        - name: date
          in: path
          required: true
          description: The date of the day
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The id of the day
                  date:
                    type: string
                    format: date
                    description: The date of the day
                  foods:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The id of the food
                        name:
                          type: string
                          description: The name of the food
                        calories:
                          type: integer
                          description: The calories of the food
                        position:
                          type: integer
                          description: The position of the order that the item should be displayed in
                  recipes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The id of the recipe
                        name:
                          type: string
                          description: The name of the recipe
                        position:
                          type: string
                          description: The position of the order that the recipe should be displayed in
                        foods:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                                description: The id of the food
                              name:
                                type: string
                                description: The name of the food
                              calories:
                                type: integer
                                description: The calories of the food
                              position:
                                type: integer
                                description: The position of the order that the item should be displayed in
        '401':
          description: Unauthorized
        '404':
          description: User or day does not exist
...
  /days:
    post:
      summary: Add a new day
      description: Add a new day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                  description: The date of the day
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The id of the new day
                  date:
                    type: string
                    format: date
                    description: The date of the new day
        '401':
          description: Unauthorized
        '400':
          description: Invalid request
  /days/{date}:
    get:
      summary: Get a day by id
      description: Get a single day
      parameters:
        - name: date
          in: path
          required: true
          description: The id of the day to get
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The id of the day
                  date:
                    type: string
                    format: date
                    description: The date of the day
                  totalCalories:
                    type: number
                    description: Total number of calories for the day, in kcal
                  totalCarbs:
                    type: number
                    description: Total amount of carbs for the day, in grams
                  totalProtein:
                    type: number
                    description: Total amount of protein for the day, in grams
                  totalFat:
                    type: number
                    description: Total amount of fat for the day, in grams
                  foods:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The id of the food
                        name:
                          type: string
                          description: The name of the food
                        calories:
                          type: number
                          description: The calories of the food
                        amount:
                          type: integer
                          description: The number of food items
                        protein:
                          type: number
                          description: Amount of protein, in grams
                        carbs:
                          type: number
                          description: Amount of carbs, in grams
                        fat:
                          type: number
                          description: Amount of fat, in grams
                        position:
                          type: integer
                          description: The position of the food in the day
                  recipes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The id of the recipe
                        name:
                          type: string
                          description: The name of the recipe
                        position:
                          type: integer
                          description: The position of the recipe in the day
                        foods:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                                description: The id of the food
                              name:
                                type: string
                                description: The name of the food
                              calories:
                                type: integer
                                description: The calories of the food
                              position:
                                type: integer
                                description: The position of the food in the recipe
        '401':
          description: Unauthorized
        '404':
          description: Day does not exist
    put:
      summary: Update a day by id
      description: Edit a day
      parameters:
        - name: date
          in: path
          required: true
          description: The id of the day to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                  description: The date of the day
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The id of the day
                  date:
                    type: string
                    format: date
                    description: The date of the day
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Day does not exist
    delete:
      summary: Delete a day by id
      description: Delete a day's entry
      parameters:
        - name: date
          in: path
          required: true
          description: The id of the day to delete
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Day does not exist
...